+++ FileIO/MeshIO/LegacyVtkInterface.cpp
@@ -168,7 +168,7 @@ void LegacyVtkInterface::WriteVTKCellData(fstream &vtk_file) const
 				break;
 		}
 
-		for(int j = 0; j < ele->GetNodesNumber(false); j++)
+		for(size_t j = 0; j < ele->GetNodesNumber(false); j++)
 			vtk_file << " " << ele->nodes_index[j];
 
 		vtk_file << endl;
